FIX: supervised sequence to vector mapping with RNNs was broken.
Breze models have an attribute sample_dim which is used to indicate
along which axis statistically independent samples lie. In the
case of RNNs, this is 1 in the typical setting. However in the case
of sequence to point mappings, the sample dimension is 1 for the input
and 0 for the output.

This was not respected up until now. I made sample_dim a property in that
case because the pooling used (i.e. 'sum' in contrast to None) uniquely
identifies this.